lifting directory local echo merging to its own class
